requirements.txt
amqp==2.6.1
asgiref==3.5.1
billiard==3.6.4.0
celery==4.4.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.4
django-picklefield==3.0.1
dnspython==2.2.1
eventlet==0.33.0
greenlet==1.1.2
kombu==4.6.11
prompt-toolkit==3.0.29
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==1.3.0
wcwidth==0.2.5
websocket-client==1.3.2
